adds power taker API endpoint
adds a contract_factory with tests and the endpoint passes the params
on to the contract_factory. adds a method for creating the first
power taker contract. it allows ti have a user without profile, a user with
profile or no user at all.

it does not allow to pick an existing contracting_party for the preceeding
contracts.

fixes #520