API cleanup, concrete result value object
the methods which expect an Interval.live only do not need the interval as
it carries not extra info.

the result objects are more tailormade to what the method actually returns.
both on the datasource side as well the on the next layer which uses the
datasources: current_power and power_charts