fix timezone problem for getting missing reading
this is fixed now for local setup and codeship, but it reveals a systematic
problem as the server is not handling timezone in consistent manner. there are
time operations which use the system timezone and others from ActiveSupport which
uses the Rails timezone.

all kind of sideeffects are possible when dealing with readings, DB and discovergy
(they use utc unix epoch times)