remove unused rack-timeout
as long some discovergy requests are taking longer then 30 seconds we get
a heroku timeout before first. but heroku does not kill the process and we
are able to finish the discovergy request and produce an entry in the cache
for this request. i.e. the next request from the client will be served from
cache.